Refactor CellStyleCreator for more efficient locale usage
Locale object 'usLocale' creation was moved up in the class hierarchy for improved code readability and efficiency. In the previous implementation, the 'usLocale' object was created every time the 'createDateCellStyle' function was called, leading to wastage of computation resources. Now, one instance is created once and reused throughout the class, improving overall performance.